backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.14
langgraph==0.2.20
ragas==0.2.3
# faiss-cpu pinned to a recent available wheel. If you are on macOS/Apple Silicon
# and this still fails, consider installing via conda-forge or building from source.
faiss-cpu==1.12.0
sentence-transformers==3.0.1
tenacity==8.5.0
python-multipart==0.0.9
websockets==12.0
pytest==8.3.2
transformers==4.44.2
huggingface_hub==0.24.6
accelerate==0.34.2
torch==2.8.0
